Replace unsigned long with size_t for sizes.
Whenever usnigned long is used for a variable in memory, use size_t.

+++ builtin/describe.c
@@ -209,11 +209,11 @@ static int compare_pt(const void *a_, const void *b_)
 	return 0;
 }
 
-static unsigned long finish_depth_computation(
+static size_t finish_depth_computation(
 	struct commit_list **list,
 	struct possible_tag *best)
 {
-	unsigned long seen_commits = 0;
+	size_t seen_commits = 0;
 	while (*list) {
 		struct commit *c = pop_commit(list);
 		struct commit_list *parents = c->parents;
