+++ parse-options.c
@@ -180,11 +180,11 @@ static int get_value(struct parse_opt_ctx_t *p,
 
 	case OPTION_MAGNITUDE:
 		if (unset) {
-			*(unsigned long *)opt->value = 0;
+			*(size_t *)opt->value = 0;
 			return 0;
 		}
 		if (opt->flags & PARSE_OPT_OPTARG && !p->opt) {
-			*(unsigned long *)opt->value = opt->defval;
+			*(size_t *)opt->value = opt->defval;
 			return 0;
 		}
 		if (get_arg(p, opt, flags, &arg))
